BigDataProject/Dynamic/requirements.txt
certifi==2023.11.17
elastic-transport==8.11.0
elasticsearch==8.11.1
findspark==2.0.1
kafka-python==2.0.2
numpy==1.26.2
pandas==2.1.4
py4j==0.10.9.7
pyspark==3.5.0
python-dateutil==2.8.2
pytz==2023.3.post1
six==1.16.0
tzdata==2023.4
urllib3==2.1.0
